edit profileDtos type mismatch

+++ ArenaBackend/ArenaApplication/Dtos/ProfileDtos/GetProfileDto.cs
@@ -16,9 +16,9 @@ namespace ArenaApplication.Dtos.ProfileDtos
         public string PreferredLanguage { get; set; } = null!;
         public bool IsActive { get; set; }
 
-        public double? Weight { get; set; }
-        public double? Height { get; set; }
-        public double? BMI { get; set; }
+        public decimal? Weight { get; set; }
+        public decimal? Height { get; set; }
+        public decimal? BMI { get; set; }
         public string? Gender { get; set; }
         public string? ProfileImage { get; set; }
         public DateOnly? Birthday { get; set; }
